Widgets should have an interface for templates
Move shakespeare specific stuff to Yesod.Shakespeare
I have not bothered to make Yesod.Shakespeare compile yet

The dependency chain is now

Yesod.Core -> Yesod.Widget -> Yesod.Shakespeare
